Fractional position logging

+++ src/dev/heliosares/auxprotect/adapters/BungeeSenderAdapter.java
@@ -63,7 +63,7 @@ public class BungeeSenderAdapter extends SenderAdapter {
     }
 
     @Override
-    public void teleport(String world, int x, int y, int z, int pitch, int yaw)
+    public void teleport(String world, double x, double y, double z, int pitch, int yaw)
             throws NullPointerException, UnsupportedOperationException {
         throw new UnsupportedOperationException();
     }
